% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamDiversity.R
\name{BamDiversity}
\alias{BamDiversity}
\title{BamDiversity
UNDER DEV
K-mer diversity of mapped reads.
returns data frame containing 3 columns
\enumerate{
\item idx     : alignment starting position of the reference sequence
\item Sequence: Nucleotide sequence including INDEL.
\item count   : # of sequence
}}
\usage{
BamDiversity(param, kmer, chrom)
}
\arguments{
\item{kmer}{(required) kmer length}

\item{chrom}{(required) sequence ids for test.}

\item{bam}{(required) bam file to read}

\item{clip}{including clip or not}
}
\value{
list or vector object.
}
\description{
BamDiversity
UNDER DEV
K-mer diversity of mapped reads.
returns data frame containing 3 columns
\enumerate{
\item idx     : alignment starting position of the reference sequence
\item Sequence: Nucleotide sequence including INDEL.
\item count   : # of sequence
}
}
