% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestblasthit.R
\name{bestblasthit}
\alias{bestblasthit}
\title{bestblasthit}
\usage{
bestblasthit(x, evalue)
}
\arguments{
\item{x}{data.frame}

\item{name_evalue}{e-value column name.}

\item{name_identity}{sequence identity column.}

\item{cutoff_evalue}{e-value threshold. any result more than the value was removed.}

\item{cutoff_identity}{identity cutoff.}
}
\value{
sorted, and best matched data.frame
}
\description{
this function sort evalue, identity, and find best result.
}
