% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioLinkGen.R
\name{bioLinkGen}
\alias{bioLinkGen}
\title{bioLinkGen}
\usage{
bioLinkGen(x, format = "text", target = "_blank", class = "auto", desc = FALSE)
}
\arguments{
\item{x}{a vector file.}

\item{format}{text or html a tag.}

\item{desc}{TRUE or FALSE. describe the accession.}

\item{when}{format is "html", 4 target link action works. (_self, _blank, _parent, or _top). please see https://www.w3schools.com/tags/att_a_target.asp}
}
\value{
list object which contains summary and reports for denoted qc reports.
}
\description{
this function generates links (bioproject, run, experiment, sample information for) for a given strings.
}
\examples{
bioLinkGen("SRR17041298")
browseURL(bioLinkGen("SRR17041298"))
}
