% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWTcode.R
\name{BWTcode}
\alias{BWTcode}
\title{BWTcode}
\usage{
BWTcode(str = "ACACT", rc = TRUE)
}
\arguments{
\item{str}{a string or vector.}

\item{rc}{if you want to count repetitive motifs from the}
}
\value{
downloaded file
}
\description{
This function is introduced from Burrow-wheeler Transform to count repetitive motifs correctly.
the development aim
}
\details{
please use vBWTcode
}
\examples{
download_SRA("SRR17041298")
}
