% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mArray.R
\name{clock_gen}
\alias{clock_gen}
\title{generate dynamic methylation clock using glmnet package.}
\usage{
clock_gen(
  x,
  y = Age,
  marker,
  test,
  offsets = 0.05205479,
  SID = SID,
  useFemale = F,
  s = c("lambda.min", "lambda.1se"),
  alpha = 1,
  weight = 1
)
}
\arguments{
\item{x}{tibble object which has n x m including SID, Age}

\item{y}{Age column name}

\item{marker}{1 dimentional character vector containing markers which passed the Quality Control}

\item{test}{test set array}

\item{useFemale}{T/F value}

\item{s}{lambda.min or lambda.1se}

\item{alpha}{1}

\item{weight}{weight}

\item{offset}{Gastrulation time. The value is used to construct aging value.}
}
\description{
generate dynamic methylation clock using glmnet package.
}
