% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{download_GEO}
\alias{download_GEO}
\title{download GEO matrix from GEO ftp.}
\usage{
download_GEO(GSE, prefix = "GEO", ...)
}
\arguments{
\item{GSE}{list of GSEs for downloading}

\item{prefix}{base directory}

\item{type}{file type to download}
}
\description{
download GEO matrix from GEO ftp.
}
\examples{
gexpit(runif(100,0,1))
}
\author{
Byungju Kim (byungju@bu.edu)
}
