% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_SRA.R
\name{download_SRA}
\alias{download_SRA}
\title{download_SRA}
\usage{
download_SRA(x, dest = ".//", credential = "no-sign-request", ...)
}
\arguments{
\item{x}{a SRA Run file accession.}

\item{dest}{target forder to download. if the folder is not exists, the function will make it.}
}
\value{
downloaded file
}
\description{
this function downloads fastq file using aws client into target folder, decrypt downloaded file into fastq files, and then remove the original file
aws client and fasterq-dump in sra-toolkit are required to run this command.
}
\details{
please see the following links to download required packages.
aws client: https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html
fasterq-dump: https://github.com/ncbi/sra-tools
}
\examples{
download_SRA("SRR17041298")
}
