% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_uniprot.R
\name{download_uniprot}
\alias{download_uniprot}
\title{download data}
\usage{
download_uniprot(x, dataset = "entry", format = "txt")
}
\description{
Redirect texts into gbff multiline outputs in metadata section
}
\details{
param x dataset accession number
param format file format
}
