% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efetch.R
\name{efetch}
\alias{efetch}
\title{efetch
UNDER DEV
search NCBI db using esearch API. this function is under development
the function returns
Returns formatted data records for a list of input UIDs
Returns formatted data records for a set of UIDs stored on the Entrez History server}
\usage{
efetch(
  id = NA,
  input = NA,
  query_key,
  WebEnv,
  db = "pubmed",
  format = c("docsum", "full", "uilist", "summary", "gene_table", "gb", "gbc", "ft",
    "gbwithparts", "fasta_cds_na", "fasta_cds_aa", "gp", "gpc", "ipg"),
  retmode = c("xml", "text", "asn.1")
)
}
\arguments{
\item{id}{(required) Unique identifier or accession number. NCBI uid. vector object.}

\item{input}{Read identifier(s) from file (one line).}

\item{db}{(required) Database name. NCBI DBs were described in details.}

\item{retmode}{(optional)  determine if the output is XML/plain text. the option will override user input if there is only one option is available.}

\item{api_key}{(optional) NCBI api_key. you can get your own api_key from NCBI setting page after log-in. To use the function easily, you need to declare "NCBI_API_KEY" in .Renviron file.}

\item{usehistory}{T/F. save result in your search history. you can also get WebEnv information. (default=T)}

\item{retstart}{(optional) number of skipping from the search  (Default=0)}

\item{retmax}{(optional) total number of unique identifiers from the retrived set to be shown in the output (default=20)}

\item{rettype}{(optional)  (optional) "uilist" or "count" (default="uilist"). if the value is count, the function only redirects the total number of query results.}

\item{datetype}{(optional) one from c("crdt","edat","pdat","mhda")}
}
\value{
list or vector object.
}
\description{
efetch
UNDER DEV
search NCBI db using esearch API. this function is under development
the function returns
Returns formatted data records for a list of input UIDs
Returns formatted data records for a set of UIDs stored on the Entrez History server
}
\details{
\tabular{llll}{
   Databases \tab Record Type \tab retmode \tab rettype \cr
   All \tab Document Summary \tab xml \tab docsum \cr
   All \tab List of UIDs in XML \tab xml \tab uilist \cr
   All \tab List of UIDs in plain text \tab text \tab uilist \cr
   MeSH \tab Full record in plain text \tab text \tab full \cr
   NLM Catalog \tab Full record in plain text \tab text \tab  \cr
   NLM Catalog \tab XML \tab xml \tab  \cr
   PubMed \tab ASN.1 (text) \tab asn.1 \tab  \cr
   PubMed \tab PubMed XML \tab xml \tab  \cr
   PubMed \tab MEDLINE (text) \tab text \tab medline \cr
   PubMed \tab PMID list (text) \tab text \tab uilist \cr
   PubMed \tab Abstract (text) \tab text \tab abstract \cr
   PubMed \tab Summary (text) \tab text \tab docsum \cr
}
\tabular{llll}{
   db \tab format \tab mode \tab Report Type \cr
   (all) \tab  \tab  \tab  \cr
    \tab docsum \tab  \tab DocumentSummarySet XML \cr
    \tab docsum \tab json \tab DocumentSummarySet JSON. \cr
    \tab full \tab  \tab Same as native except for mesh \cr
    \tab uid \tab  \tab Unique Identifier List \cr
    \tab url \tab  \tab Entrez URL \cr
    \tab xml \tab  \tab Same as -format full -mode xml \cr
    \tab  \tab  \tab  \cr
   bioproject \tab  \tab  \tab  \cr
    \tab native \tab  \tab BioProject Report \cr
    \tab native \tab xml \tab RecordSet XML. \cr
    \tab  \tab  \tab  \cr
   biosample \tab  \tab  \tab  \cr
    \tab native \tab  \tab BioSample Report \cr
    \tab native \tab xml \tab BioSampleSet XML \cr
   biosystems \tab  \tab  \tab  \cr
    \tab native \tab xml \tab Sys-set XML \cr
    \tab  \tab  \tab  \cr
   clinvar \tab  \tab  \tab  \cr
    \tab variation \tab  \tab Older Format \cr
    \tab variationid \tab  \tab Transition Format \cr
    \tab vcv \tab  \tab VCV Report \cr
    \tab clinvarset \tab  \tab RCV Report \cr
   gds \tab  \tab  \tab  \cr
    \tab native \tab xml \tab RecordSet XML \cr
    \tab summary \tab  \tab Summary \cr
   gene \tab  \tab  \tab  \cr
    \tab full_report \tab  \tab Detailed Report \cr
    \tab gene_table \tab  \tab Gene Table \cr
    \tab native \tab  \tab Gene Report \cr
    \tab native \tab asn.1 \tab Entrezgene ASN.1 \cr
    \tab native \tab xml \tab Entrezgene-Set XML \cr
    \tab tabular \tab  \tab Tabular Report \cr
   homologene \tab  \tab  \tab  \cr
    \tab alignmentscores \tab  \tab Alignment Scores \cr
    \tab fasta \tab  \tab FASTA \cr
    \tab homologene \tab  \tab Homologene Report \cr
    \tab native \tab  \tab Homologene List \cr
    \tab native \tab asn.1 \tab HG-Entry ASN.1 \cr
    \tab native \tab xml \tab Entrez-Homologene-Set XML \cr
    \tab  \tab  \tab  \cr
   mesh \tab  \tab  \tab  \cr
   full                        Full Record \tab  \tab  \tab  \cr
   native                      MeSH Report \tab  \tab  \tab  \cr
   native             xml      RecordSet XML \tab  \tab  \tab  \cr
}


|
| nlmcatalog
|                 native                      Full Record
|                  native             xml      NLMCatalogRecordSet XML
|
| pmc
|                bioc                        PubTator Central BioC XML
|                medline                     MEDLINE
|                native             xml      pmc-articleset XML
|
| pubmed
|                abstract                    Abstract
|                bioc                        PubTator Central BioC XML
|                medline                     MEDLINE
|                native             asn.1    Pubmed-entry ASN.1
|                native             xml      PubmedArticleSet XML
|
| (sequences)
|                acc                         Accession Number
|                est                         EST Report
|                fasta                       FASTA
|                fasta              xml      TinySeq XML
|                fasta_cds_aa                FASTA of CDS Products
|                fasta_cds_na                FASTA of Coding Regions
|                ft                          Feature Table
|                gb                          GenBank Flatfile
|                gb                 xml      GBSet XML
|                gbc                xml      INSDSet XML
|                gene_fasta                  FASTA of Gene
|                gp                          GenPept Flatfile
|                gp                 xml      GBSet XML
|                gpc                xml      INSDSet XML
|                gss                         GSS Report
|                ipg                         Identical Protein Report
|                ipg                xml      IPGReportSet XML
|                native             text     Seq-entry ASN.1
|                native             xml      Bioseq-set XML
|                seqid                       Seq-id ASN.1
| snp
|                  json                        Reference SNP Report
| sra
|                 native             xml      EXPERIMENT_PACKAGE_SET XML
|                  runinfo            xml      SraRunInfo XML
| structure
|                mmdb                        Ncbi-mime-asn1 strucseq ASN.1
|                  native                      MMDB Report
|                  native             xml      RecordSet XML
|
| taxonomy
|                  native                       Taxonomy List
|                  native             xml       TaxaSet XML
}
