% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einfo.R
\name{einfo}
\alias{einfo}
\title{einfo}
\usage{
einfo(db = "")
}
\arguments{
\item{db}{database name. if this field is "", it shows all of the available databases.}
}
\description{
Provides the number of records indexed in each field of a given database, the date of the last update of the database, and the available links from the database to other Entrez databases.
NCBI E-utility wrapper series for R users.
}
\details{
https://www.ncbi.nlm.nih.gov/books/NBK25497/
}
