% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esearch.R
\name{esearch}
\alias{esearch}
\title{esearch}
\usage{
esearch(
  query,
  db = "pubmed",
  api_key = Sys.getenv("NCBI_API_KEY"),
  usehistory = TRUE,
  retmax = 20,
  retstart = 0,
  rettype = "uilist"
)
}
\arguments{
\item{query}{NCBI string query. for more information please read}

\item{db}{NCBI DBs described in details}

\item{api_key}{NCBI api_key. you can get your own api_key from NCBI setting page after log-in. To use the function easily, you need to declare "NCBI_API_KEY" in .Renviron file.}

\item{usehistory}{T/F. save result in your search history. you can also get WebEnv information. (default=T)}

\item{retmax}{total number of unique identifiers from the retrived set to be shown in the output (default=20)}

\item{retstart}{number of skipping from the search  (Default=0)}

\item{rettype}{(optional) "uilist" or "count" (default="uilist"). if the value is count, the function only redirects the total number of query results.}

\item{datetype}{(optional) one from c("crdt","edat","pdat","mhda")}
}
\value{
list or vector object.
}
\description{
search NCBI db using esearch API. retun
}
\details{
available DBs :
\itemize{
\item https://www.ncbi.nlm.nih.gov/books/NBK25497/table/chapter2.T._entrez_unique_identifiers_ui/?report=objectonly
\item or use the function einfo()
}

Followings are examples of search field tags from https://pubmed.ncbi.nlm.nih.gov/help/#proximity-searching.
You may also get the information by \code{einfo(db="pubmed")$Field}\tabular{lll}{
   Affiliation \link{ad} \tab All Fields \link{all} \tab Article Identifier \link{aid} \cr
   Author \link{au} \tab Author Identifier \link{auid} \tab Last Author Name \link{lastau} \cr
   PMID \link{pmid} \tab Title/Abstract \link{tiab} \tab Grant Number \link{gr} \cr
   Publication Date \link{dp} \tab Title \link{ti} \tab MeSH Terms \link{mh} \cr
   Text Words \link{tw} \tab Title \link{ti} \tab MeSH Major Topic \link{majr} \cr
}
}
