% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{expit}
\alias{expit}
\title{expit transformation}
\usage{
expit(x)
}
\arguments{
\item{x}{data}
}
\value{
logistic transformation of x
}
\description{
\code{expit} returns logistic transformation of a given data x
}
\examples{
gexpit(runif(100,0,1))
}
\author{
Byungju Kim (byungju@bu.edu)
}
