% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extVct.R
\name{extVct}
\alias{extVct}
\title{extVct}
\usage{
extVct(x, y, k)
}
\arguments{
\item{x}{index. numeric value}

\item{y}{chracter vector to manipulate.}
}
\value{
vector of characters. \code{y[x:length(y)-(k-x)]}
}
\description{
return character vector ranging index x to k-x for given vector. this function is developed to k-mer extraction for a given vector.
}
\examples{
testvector<-sample(letters,1000,replace=T)
extVct(1,testvector,10)[1:10]
extVct(2,testvector,10)[1:10]
}
