% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbff_linebreaker.R
\name{gbff_linebreaker}
\alias{gbff_linebreaker}
\title{gbff_linebreaker()}
\usage{
gbff_linebreaker(vc, sep = " ", linelimit = 67, header = "DEFINITION")
}
\description{
Redirect texts into gbff multiline outputs in metadata section
}
\details{
param
}
