% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_genes.R
\name{get_direct_children}
\alias{get_direct_children}
\title{get GO_IDs from full}
\usage{
get_direct_children(go_id, ontology = "CC")
}
\arguments{
\item{go_id}{문자열(Character). (예: "GO:0002376").}

\item{ontology}{CC or BP or MF}
}
\description{
get GO_IDs from full
}
