% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_ranked_taxonomy.R
\name{get_full_ranked_taxonomy}
\alias{get_full_ranked_taxonomy}
\title{get_full_ranked_taxonomy}
\usage{
get_full_ranked_taxonomy()
}
\arguments{
\item{locus}{(required)}
}
\description{
download and read get_full_ranked_taxonomy
}
