% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GO_genes.R
\name{get_GO_genes}
\alias{get_GO_genes}
\title{get GO_IDs from full}
\usage{
get_GO_genes(
  go_id,
  org_db = org.Mm.eg.db,
  evidence_filter = c("EXP", "IDA", "IEA")
)
}
\arguments{
\item{go_id}{문자열(Character). (예: "GO:0002376").}

\item{org_db}{dbname. (e.g:org.Mm.eg.db - mouse db)}

\item{evidence_filter}{= c("EXP","IDA","IEA)}
}
\description{
get GO_IDs from full
}
