% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEO.R
\name{getGSEdir}
\alias{getGSEdir}
\title{get full PATH of a given GSE record.}
\usage{
getGSEdir(GSE, prefix = "data/GEO")
}
\arguments{
\item{GSE}{single GSE accession}

\item{prefix}{base directory which have GSEs.}
}
\value{
single vector of PATH
}
\description{
parse GSE accession and build GSE directory.
}
\author{
KIM, BYUNGJU
}
