% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKmer.R
\name{getKmer}
\alias{getKmer}
\title{getKmer}
\usage{
getKmer(x, k = 18, count = T, norm = T)
}
\arguments{
\item{x}{DNA or character vector}

\item{k}{k-mer}
}
\value{
vector of values for given xpath.
}
\description{
This function is R version feature counting for FFP by vector rotating.
}
\examples{
xml_doc<-xml2::read_xml("<animals><pet>Cat</pet><pet>Dog</pet><insect>Ant</insect><insect>Bee</insect></animals>")
xml2vector(xml_doc,"//pet")
xml2vector(xml_doc,"//insect")
}
