% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gffcmpr.R
\name{gffcmpr}
\alias{gffcmpr}
\title{gffcmpr}
\usage{
gffcmpr(x, y)
}
\description{
The function compares two gff list loaded from read_gff function in BKbiokit
This function is desgined to compare two or more \code{liftover} or \verb{annotation result}
@param x list object loaded from read_gff
@param y list object loaded from read_gff
@return merged list output with some notation.
@examples
}
