% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{data}
}
\value{
inverse of logistic transformation of a given data x
}
\description{
\code{logit} returns inverse of logistic transformation
}
\examples{
gexpit(runif(100,0,1))
}
\author{
Byungju Kim (byungju@bu.edu)
}
