% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bwt_tbl.R
\name{make_bwt_tbl}
\alias{make_bwt_tbl}
\title{construct BWA table for specific k-mer}
\usage{
make_bwt_tbl(x)
}
\arguments{
\item{x}{k-mer}

\item{format}{file format}
}
\description{
Redirect texts into gbff multiline outputs in metadata section
}
