% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{makeGEOPath}
\alias{makeGEOPath}
\title{makeGEOPath}
\usage{
makeGEOPath(GSE, type = c("matrix", "suppl", "xml", ""), prefix = "GEO", ...)
}
\arguments{
\item{GSE}{GSE accession}

\item{type}{type of file format you download}
}
\description{
makeGEOPath
}
\author{
Byungju Kim (byungju@bu.edu)
}
