% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phylip_dist.R
\name{read.phylip_dist}
\alias{read.phylip_dist}
\alias{read_phylip_dist}
\title{read_phylip_dist: read Phylip style distance matrix as distance matrix in R}
\usage{
read_phylip_dist(file, rm.list = c(), skip = 1)
}
\arguments{
\item{file}{a file name,}

\item{rm.list}{list for sample exclusion.}

\item{skip}{a numeric value of line skipping.}
}
\description{
read_phylip_dist: read Phylip style distance matrix as distance matrix in R
}
\examples{
read.dist(matrix.tsv)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
