% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastqc.R
\name{read_fastqc}
\alias{read_fastqc}
\title{read_fastqc}
\usage{
read_fastqc(fastqc = "K34_1_fastqc.zip")
}
\arguments{
\item{fastqc}{a zip file for fastqc output (FILE_fastqc.zip)}
}
\value{
list object which contains summary and reports for denoted qc reports.
}
\description{
this function read a given fastqc output (.zip files within specific directory) and saving them into
}
\examples{
read_fastqc("demo_fastqc.zip")
}
