% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastqcdir.R
\name{read_fastqcdir}
\alias{read_fastqcdir}
\title{read_fastqcdir}
\usage{
read_fastqcdir(DIR)
}
\arguments{
\item{DIR}{a directory which stores fastqc results.}
}
\value{
list object which contains aggregated all fastqc reports within the directory.
}
\description{
this function read all fastqc output (.zip files within specific directory) and saving them into
}
\examples{
read_fastqcdir(DIR)
}
