% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{read_geov2}
\alias{read_geov2}
\title{read_geo}
\usage{
read_geov2(GSE, prefix = "GEO", ...)
}
\arguments{
\item{GSE}{GSE accession}

\item{prefix}{the directory which you saved all of GSE structure.}
}
\value{
inverse of logistic transformation of a given data x
}
\description{
\code{logit} returns inverse of logistic transformation
}
\author{
Byungju Kim (byungju@bu.edu)
}
