% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff.R
\name{read_gff}
\alias{read_gff}
\title{read_gff}
\usage{
read_gff(x, comment = "#")
}
\arguments{
\item{x}{gff file}
}
\value{
data frame which contained cleanned
}
\description{
read gff file into wide-form tibble Object.
}
