% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\name{read_gmt}
\alias{read_gmt}
\alias{read_GMT}
\alias{write_gmt}
\title{Read/Write MySigDb gmt file format}
\usage{
read_GMT(x, simple = TRUE)

write_gmt(x, file, url = "BKbiokit")
}
\arguments{
\item{x}{list file.}

\item{file}{output file}

\item{url}{url}
}
\value{
list object which containing all gene sets in a given gmt file.

saving file.
}
\description{
Read GMT files into list object
}
\author{
KIM, BYUNGJU
}
