% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEO.R
\name{read_GSE}
\alias{read_GSE}
\title{read_GSE}
\usage{
read_GSE(
  GSE,
  prefix = "GEO",
  keep = c("Series_title", "Series_summary", "Series_overall_design", "Series_type"),
  ...
)
}
\arguments{
\item{GSE}{GSE accession}

\item{prefix}{base directory}

\item{keep}{vector of categories you want to keep in the returned object.}
}
\value{
list object containing description about the GSE accession
}
\description{
read GSE documents downloaded by BJK
}
