% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEO.R
\name{read_GSM}
\alias{read_GSM}
\title{read_GSM}
\usage{
read_GSM(GSE, prefix = "GEO", ...)
}
\arguments{
\item{GSE}{GSE accession}

\item{prefix}{base directory. We read data from prefix/GSEXXXnnn/GSEXXXXXX}
}
\description{
read_GSM
}
\author{
KIM, BYUNGJU
}
