% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trimmomatic}
\alias{trimmomatic}
\title{trimmomatic}
\usage{
trimmomatic(
  x,
  ext = c("_1.fastq.gz", "_2.fastq.gz"),
  out1 = c("_1.trimmed.fastq.gz", "_1.unpaired_trimmed.fastq.gz"),
  out2 = c("_2.trimmed.fastq.gz", "_2.unpaired_trimmed.fastq.gz"),
  threads = 1,
  logfile = "",
  summary = "",
  phred = 33,
  AdapFile = "",
  AdapAllowMaxMM = 2,
  AdapPEQ = 30,
  AdapSEQ = 10,
  leading = 15,
  trailing = 15,
  command = ""
)
}
\arguments{
\item{x}{sample fastq file basename}

\item{ext}{extension for fastq file. if the length of ext vector is 1, we consider it as single-end reads.}
}
\description{
Trimmomatic: A flexible read trimming tool for Illumina NGS data
}
