% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vct_lookup.R
\name{vct_lookup}
\alias{vct_lookup}
\alias{\%lookup\%}
\title{vector lookup utility}
\usage{
vct_lookup(query, ref)

query \%lookup\% ref
}
\arguments{
\item{query}{A vector to be mapped (Left Hand Side for the operator).}

\item{ref}{A named vector serving as the lookup dictionary (Right Hand Side for the operator).
The names of the vector are the keys, and the values are the replacements.}
}
\value{
A vector of the same class as \code{query} with mapped values.
}
\description{
Maps elements of a vector using a named vector.
Elements present in the dictionary keys are replaced with their corresponding values.
Elements not found in the dictionary remain unchanged.
}
\examples{
# 1. Define data
my_genes <- c("Gnas1", "Tp53", "Unknown")
conversion_table <- c("Gnas1" = "Gnas", "Tp53" = "TP53")

# 2. Use the standard function
vct_lookup(my_genes, conversion_table)

# 3. Use the infix operator
my_genes \%map_by\% conversion_table

}
