% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ORA.R
\name{write_ora_results}
\alias{write_ora_results}
\title{Write all ORA tables with a common prefix, without overwriting by default.}
\usage{
write_ora_results(x, output_prefix)
}
\description{
Write all ORA tables with a common prefix, without overwriting by default.
}
