% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2vector.R
\name{xml2vector}
\alias{xml2vector}
\title{xml2vector}
\usage{
xml2vector(xml, path)
}
\arguments{
\item{xml}{xml object}

\item{path}{xpath}
}
\value{
vector of values for given xpath.
}
\description{
this function extract spacific path(xpath) node values and convert them into vectors
}
\examples{
xml_doc<-xml2::read_xml("<animals><pet>Cat</pet><pet>Dog</pet><insect>Ant</insect><insect>Bee</insect></animals>")
xml2vector(xml_doc,"//pet")
xml2vector(xml_doc,"//insect")
}
