% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{categorize the input column}
\usage{
categorize(
  data,
  input_col,
  output_col,
  conditions,
  lvl = NULL,
  default = NA_character_
)
}
\arguments{
\item{data}{data frame}

\item{input_col}{input column}

\item{output_col}{output column (categorized column)}

\item{lvl}{level}
}
\description{
The function categorize the column based on the colum value and. condition/rule.
}
