% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCm38_to_GRCm39}
\alias{GRCm38_to_GRCm39}
\title{UCSC GRCm38 to mm10 gene symbol changes}
\format{
A data frame with 64 rows and 3 variables
\describe{
\item{GRCm38}{UCSC/igenome gene symbol}
\item{GRCm39}{mm10 gene symbol}
}
}
\usage{
GRCm38_to_GRCm39
}
\description{
A dataset containing triplet codon(codon), 1-letter amino acids (AA1), 3-letter amino acids (AA3), and property (property)
}
\keyword{datasets}
