% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff.R
\name{read_gff}
\alias{read_gff}
\title{read_gff}
\usage{
read_gff(x, comment = "#")
}
\arguments{
\item{x}{gff file}
}
\value{
data frame which contained cleanned
}
\description{
\code{read_gff} converts gff file into list of wide-form tibble Objects based on features.
}
\details{
for the file format of gff, please see https://www.ensembl.org/info/website/upload/gff.html?redirect=no
}
