% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_genetic_code}
\alias{standard_genetic_code}
\title{Standard genetic table}
\format{
A data frame with 64 rows and 3 variables
\describe{
\item{codon}{triplet codon AAA to TTT}
\item{AA1}{1 letter amino acid code}
\item{AA3}{3 letter amino acid code}
\item{property}{chemical property}
}
}
\usage{
standard_genetic_code
}
\description{
A dataset containing triplet codon(codon), 1-letter amino acids (AA1), 3-letter amino acids (AA3), and property (property)
}
\keyword{datasets}
